Reimplement constitutive output maps
Previously, the constitutive output maps were maps of raw pointers,
with the data attached to each element. To decouple this, and enable
more complex constitutive laws (e.g. parallel and in series rheological
models), the maps now contain unique pointers, i.e. the data now "lives
and dies" with the map.

Most of the changes are to the constitutive laws and the elements, to
go along with the changed interface.

A lot of changes were needet for the AdditiveOutput and AdditiveInput
laws. I've added two unit tests, but this is still far from stable.