Add test files for coupled thermomechanics
I've added an example of a heated plate, with nonlinear thermal
expansion coefficients (`HeatedPlate.cpp`). It also acts as a small test
file (`ThermoMechanics2D.cpp`), which is the same example with a smaller
mesh. But it only checks if it runs/converges, not whether the results
are correct.

Also added checking the Hessian1 and internal gradient for heat
conduction in the ContinuumElement unit test.