Remove EvaluateData object from elements
The EvaluateData objects contained mostly constitutive data that was
pointed to by the constitutive input/output maps. With the new
implementation of these maps, these objects became mostly redundent. The
remaining data was moved directly to the element, and many of the
function calls are now a bit cleaner.

Also, I've encountered several places in the ContinuumBoundaryElement
that I had not properly converted to the new maps. As a result, the
MoistureTransport test passes again.