Partially revert commit 9545875
Getting rid of the EvaluateData completely was a bit rash. Storing the
shape functions and nodal values at the element permanently will
increase memory usage. Therefore, EvaluateData is back. I have removed
it

Naturally, the fix for the MoistureTransport test remains. Also, I
forgot to add the CMakeLists for the constitutive tests.