Change syntax for remaining tests
Everything now compiles, but three tests don't run. Specifically,
two more changes are needed: The moisture transport law constructor gets
a control node, but nowhere is this being set. In the
MultipleConstitutiveLaw and the Shrinkage tests, two different
"constitutive law trees" are being assembled, but static data is trying
to be accessed the same way. This can't work and needs to be adjusted.