Resurrect examples
I resurrected the examples; mostly by changing the enums that had
changed in NuTo.

Also, there is now a .clang-format file, which you can use to format
your code with. We should agree on a common style. What I have checked
in is intented as a baseline and open for discussion.

Also, Travis should now make a release build as well.