Remove ENABLE_DEBUG option
Debug/Release is not really a project option, but a build option. By
setting it with ENABLE_DEBUG, we restricted ourselves to binary Debug
or Release. The convention with cmake is to do

    cmake -DCMAKE_BUILD_TYPE=Debug ../nuto

if you want a debug build, and `Release` for, you guessed it, a release
build. But now we can also say `RelWithDebInfo`, i.e.  with
optimizations, but including the debugging symbols. Also, there is one
less option; we have to many already.

The macro `DEBUG` is gone along with the option. The standard uses
`NDEBUG` to disable asserts; this is what you should use. So simpley replace
`#ifdef DEBUG` with `#ifndef NDEBUG`.

FYI: By default, cmake does "nothing". So if you don't set the
CMAKE_BUILD_TYPE, you'll get neither debugging symbols nor
optimizations.