Fixes doxygen documentation header
We used custom header, footer, and style sheet files defined in
Doxyfile.in:

HTML_HEADER
HTML_FOOTER
HTML_EXTRA_STYLESHEET

I suspect that some of the links in header.html are not working anymore
and thus caused the documentation to not display properly.

I removed the custom files from Doxyfile.in.