Seperate Python wrapper code from C++ sources
The Python wrapper was too intertwined with the rest of the code. Building NuTo
necessarily meant building the Python wrappers and the same CMakeLists.txt of a
module had to deal with C++ and Python at the same time.

Now, there is an option, ENABLE_PYTHON, with which you can decide if you want
to build the wrapper. All the SWIG files and the CMake stuff necessary for the
have moved to a seperate directory, `python/`.