Reenable the ParallelAssembly test; sort of
The test had been disabled because it failed when run on machines with
high load. I've modified it so that it compiles again and reenabled it
in cmake, but not in the tests. I.e. it is compiled, but not run. Seems
like a reasonable compromise until we have a better solution.