Introduce "object libraries" in cmake
Before this commit, every cpp file that had a unit test was build at
least twice: once for the lib, and once for the unit test. This commit
introduces a bit of cmake magic (in `scripts/NuToMacros.cmake`) to
generate an "object library" (cmake lingo) for each source file.

Now we link to these objects libraries and avoid duplicate compiles.