Clear `all_unit_tests` variable each CMake run
The names of all the unit tests were always appended to the list, so if
you ran CMake five times, each unit test would appear five times in the
list. More problematically, if you renamed a unit test, CMake would look
for both the old and the new, and would then error.