Create sections with "Named Constructor Idiom"
- section enum is gone
- Structure Section stuff is gone
- Volume section is gone

Instead of using the structure as a factory and owner for the sections,
this creates the sections using the "Named Constructor idiom" (see
https://isocpp.org/wiki/faq/ctors#named-ctor-idiom). This means they now
have a static `Create` method that gives back a `shared_ptr` to the
object. This method also takes the thickness/area/whatever at
construction, rather than using `SetWhatever` later.