Export cylinder specimen as cylinder in Gmsh files
The specimen in GeometryConcrete can be of type box, dogbone or
cylinder. While the placement of spheres was within that geometry, the
resulting Gmsh file contained the bounding box. At least for cylinders,
the Gmsh geometry is now also a cylinder. At some point, someone should
add the same for dogbone specimen.