Start implementing the IntegrationType conversion
The interface from IntegrationType to visualize is not ideal. Straight
forward way is to pass a rVisualize object and the IntegrationType
defines all points and cells. Then it returns the point ids with its
local coordinates (for interpolation DOF data) and the cell ids with the
corresponding integration point id (for the IP data). Increases the
coupling though.
[ci skip]