Add google benchmark
The google benchmark repository is added as submodule in the
external directory. The new cmake script adds the directory to the
build system and calls a shell script.

The purpose of the shell script is to automatically clone the
repository and set HEAD to the latest stable release
(hardcoded V 1.3.0). The "manual clone" is necessary because
submodules are not cloned together with the main repository.
The directory would just be empty and you have to do it yourself.

Also added a simple example from the documentation.