Add new benchmark example and add_benchmark macro
There is a new benchmark named SolversGB.cpp which is the google
benchmark version of Solvers.cpp.

The add_benchmark macro adds -Wno-suggest-override to the compile
flags of each benchmark. Otherwise the included benchmark header
will generate this warning. There might be a smarter way to
disable it (declare lib as system) but I couldn't
figure it out so far.